[Defect #10755] - ImageServer Alerts cause failure after upgrade
Implemented a change in how the Alert Context field is serialized.  Previously
we were putting the assembly version for the type in the serialized xml, changed
so we have the Class, and assembly name, but not the version.  Confirmed we
can still deserialize with this.  This should resolve the prior bug.

Also made a few code cleanup changes as part of this, including making the
serialized WorkQueueAlertContextData inherit from DataContractBase.

--HG--
branch : Phoenix5