[Defect #10761] - Reconciliation issue if do not refresh list.
Updated the Study Integrity Queue to use DataKeys to keep track of the selected entries,
instead of indices in the list of results.  This resolves the problem.

Note that there's a bit of odd behavior, in that if you go back and highlight
a try and reconcile an item that you've already completed, it just resets the
list, and doesn't open the dialog.  This seems sufficient, given the workflow of
selecting multiple items in the list and going through reconciling them.

--HG--
branch : Phoenix5