[Ticket #10555] - Workstation start-up time has noticably increased
Modified the plugin cache to include a checksum based on the set of plugin candidates found in the plugins folder, and the set of config files found in the install folder. For each file, the file name, create time, and last write time are included in the checksum.

This is more robust than the previous strategy of simply ensuring that the write time of the cache file was later than any other write times, as it accounts for the cases where plugins are deleted, or plugins are added but with earlier write times.

--HG--
branch : Phoenix5