[Defect #10781] - DICOM Sequence streamed out with empty tags and fixed length encoded incorrectly
Defect fix for this ticket caused the DICOMDIR functionality to break.
Discovered a couple issues with how the DICOMDIR code was trying to calculate
the length of the DICOMDIR that were hidden by this bug.  One, the DICOMDIR
saving code was not calcuating the meta info length including group length
attributes, so the group length tag was not included.  Second, this bug
caused the SQ attribute length to be include din the length, even though it
didn't have a value, which offset the mistake calculating the meta info.

Also cleaned up how the write options were used in the file, and adjusted
some comments.

--HG--
branch : Phoenix5