[Requirement #10381] - Ensure the source code repository is managed correctly
Did some more cleanup with the About.aspx page in the open source version.
Fixed the product settings to have proper licensing.  Updated the build so it
includes the changed product settings in each of the application configuration
files.  Also fixed a small issue with a reference being missing for the open
source build.

--HG--
branch : Phoenix5