[Defect #10799] - Unecessary plugins should be removed from ImageServer web directory
As part of this ticket, did some refactoring of the ImageServer & WebPortal
builds.  Changes include:
 * Added new *.targets files to the build.  These include specific groupings of files around the Webstation, WebstationDX, Services, etc.
 * The targets files include file lists, the manifest file for the grouping, plus config files to combine.
 * Renamed some of the ImageServer's manifest input files to match the tagets file names.
 * File lists for various functionality are not shared between the WebPostBuild + the ShredHost postbuild script file, and where apporpriate with the WebPortal.

Also as part of this, removed the Webstation DX plugins that were being copied
into the Web directory as part of the ImageServer, which was the core part of this
ticket.  Decided not to completely remove the Silverlight Webstation plugins for
now, although this could be done easily by just not including the targets file.

Finally, modified the web projects so that they delete the "bin" directory
contents before every compile.  Had some issues with DLLs hanging around
incorrectly when changing builds, this should help take care of that.