[Defect #9597] - Open Source Build of ImageServer doesn't work
Fixed the msbuild project for building the ImageServer. The community edition
project had not been updated to visual studio 2010, and there were a number of
issues with directories and placement of the distribution build that were
cleaned up.

Also, some source files had a few issues that needed to be cleaned up.  Finally,
fixed a few issues with the synchronization projects to deal with the ascx
files that needed to be updated due to the lack of the webstation in the
community edition.