[Ticket #11142] - The research PACS shall be accessible via a REST API
Created an RP-specific Linq-to-Sql-based implementation of the IPersistenceContext framework that is defined in Enterprise.Core.  Not sure if this is a great idea or not (it took a bit of wrangling to adapt L2S to fit those interfaces, and it's still not a perfect fit), but the upsides are:
 * re-use of the PersistenceScope infrastructure for managing connection/transaction lifetime.
 * consistent data access patterns via implementations of IPersistenceBroker.
 * a (vague) sense of consistency with other CC products that make use of the IPersistenceContext framework.

If it turns out to be problematic, we can make other plans, but for the moment it seems to work well enough.

--HG--
branch : Rigel