[Ticket #11146] - The ImageServer shall support research partitions
Added in the tables to specify a partition as a research partition and the
ability to add alternate AE titles for a partition.  This checkin should include
all the DB updates including an upgrade script and updates to the standard
table creation scripts to include the new columns and tables.  The
InsertServerPartition stored procedure was also updated to include setting the
ServerPartitionTypeEnum value.

--HG--
branch : Rigel