[Ticket #11146] - The ImageServer shall support research partitions
Re-implemented some updates that were lost when I tried to do a "rebase".
Most of the updates were simply changes to reflect the refactoring done to
make CC.IS.Common not be dependent on CC.IS.Model.  Also had to re-do
ServerPartitionMonitor, which was the main code to rework.   Retested and looks
like the adding/removing partitions and secondary AE titles works properly.

--HG--
branch : Rigel