[Ticket #11148] - The ImageServer shall support a WCF API for configuration of research partitions
Another round of updates to the data contracts:
 * Changed the general Attributes to by PolymorphicDataContractAttribute so the Jsml serializer will work with the objects
 * Introduced a new ImageServerSerializer that can serialize/deserialize the Request/State/Notification objects
 * Introduced the NewSopNotification class which will be the notification object sent via REST when new SOPs arrive.
 * Cleaned up a silly issue with my prior commit that I didn't use the Platform.Service call properly and the service objects wouldn't have been disposed.

I'm not sure yet if I should split the attributes for the Request objects vs.
the State objects, still need to look into that.  Also there's issues with
the OperationId and how its used, and a bit of refinement needed on that and
fleshing out the request State objects so they're a bit more useful.

--HG--
branch : Rigel