[Ticket #11148] - The ImageServer shall support a WCF API for configuration of research partitions
Changed all the Async bridge calls so they dont' return the State objects, it
didn't make sense to have this since the state is intended to be sent via rest
back to the client if requested.

Implemented the first research configuration services and added in appropriate
classes to handle this.

Cleaned up FaultContracts a bit on the services.  Found a few places where
OperationIds were still mentioned, cleaned this up also.

Updated OperationId references to OperationToken where it made sense.

Added shells for processing most of the requests.  This includes factory classes
for getting the processors and the processors themselves.

--HG--
branch : Rigel