[Ticket #11148] - The ImageServer shall support a WCF API for configuration of research partitions
Added unit tests for the basic serialization of requests.  This mainly is for
the Work external requests.  Also added tests for the research configuration
processors to test basic functionality.  To make these work, had to make a
change to CC.IS.Enterprise.SqlServer to create a default connection string
when running in debug mode that just uses the identity of the current user
when connecting to the database.  Also, fixed some issues with the research
plugins not being marked as plugins.  Finally, made some changes to the XmlUtils
to support serializing the State objects which require a list of possible types
needed when serializing.

--HG--
branch : Rigel