[Ticket #11148] - The ImageServer shall support a WCF API for configuration of research partitions
Introduced new NotificationQueue, ExternalRequestQueue, and ExternalRequestQueueStatusEnum
tables into the database.  These are for asynchronous processing and for notification
requests that are input into the database.  The update and db creation scripts should
all be in place for this.

Introduced the processor for notification requests for new sops and added
inserts into the database.  Also, introduced some changes to verify requests
before they're inserted into the database.  This will apply for config & work
requests, also.

--HG--
branch : Rigel