[Ticket #11148] - The ImageServer shall support a WCF API for configuration of research partitions
Created new testapp for testing the IS web services.  Its serializes/deserializes
the request objects and allows you to edit them before sending, then saves the state
object that comes up.

As part of the dev of the test app, made changes to actually get the web services
operational in the IS.  The services are now mounted and available.  There
was sone configuration and attributes incorect on the services.  Right now
just the configuration requests are implemented, but will add more as I go along.
Also made some general improvements.

--HG--
branch : Rigel