[Ticket #11209] - Research PACS notified of arriving studies
Added in simple extension point for firing events.  Instead of registering an
event handler with a specific class, an extension point must be created to
receive the event.  When an event occurs, the EventManager will fire the vent
to all extensions that specify the specific eventargs to be processed.

The extension for handling when a new SOP is added is wired up, but hasn't
been tested yet and it still needs a bit of work..

--HG--
branch : Rigel