[Ticket #11211] - The ImageServer shall support upload and download operations through a REST API
Added new Authorization status in the database for external requests, and then
implemented that the requests where "authorization" is requested.  These are
just inserted into the database and left there with the Authorization status.

Also, did preliminary implementation of Cancel requests for the uploads.  This
requred adding a "OperationTokenToCancel" field into each of the cancel requests
and then a delete of the actual matching operational tokens.

--HG--
branch : Rigel