[Ticket #11209] - Research PACS notified of arriving studies
Added in usage of the NotificationQueue for sending messages.  We now insert into
the table and the servicelock processor actually does the send.

Also, reworked the ImageServerNotification class and redid the serialization
of the notification messages.

Finally made a change to classes so the state objects inherit from the notification
object so that serialization is easier.

--HG--
branch : Rigel