[Ticket #11209] - Research PACS notified of arriving studies
Another rework of notifications based on a bug that Jon found where the
wrong operation token could be sent in a newsopnotification if two CopyStudy
requests were made with the same destination study.  The problem was
caused by the two copy requests using the same WorkQueue entry to process
the destination study objects.

Made an update to the WorkQueueUid table to add a new WorkQueueUidData xml
data field, and inserted the operation token in it for each SOP Instance
being processed.

Also, adapted this new xml field to include some of the other
optional WorkQueueUid table entries that are only used for duplicate processing.
Note that we fill the fields in properly, but are not using them.  If I can figure
out a good way to migrate the potential old records on upgrade, I'd like to
remove the other fields from the table.  Note, also, if we figure out how to
do this, it might make it easier to change a bunch of the serialization code
in the database and clean that up on upgrade, too.

Anyways, this looks good to go.  Note that the new column must be added
into the database.

--HG--
branch : Rigel