[Ticket #11211] - The ImageServer shall support upload and download operations through a REST API
Updates to upload and copy process.  The upload now supports the various options
for reassigning UIDs, and it should work to keep UIDs consistant across multiple
REST posts of the upload for a given upload request.

As part of the upload, also changed the Copy processor to use the new
DicomUidUpdator class to replace UIDs where appropriate.

Also made some updates to get true "progress" information for the Copy state
object.

--HG--
branch : Rigel