[Ticket #11250] The ImageServer shall support deletion operations through the WCF API
Added initial deletion functionality to delete Study, Series, or specific
Instances.  In general the code is functional, however, it does not as yet
properly handle if the last Sop Instance in a series is deleted or if the only
series/instance in a study is deleted.  THis needs some work yet.

Also, did some reworking of the RemoveInstanceFromStudyXmlCommand to be an
Aggregate command and to take advantage of the existing code to
write a studyxml file and being able to revert the changes.

--HG--
branch : Rigel