[Ticket #11211] - The ImageServer shall support upload and download operations through a REST API
Added config option to set the REST base URL.  Also added proper check to make
sure a file is deidentified if requested in the upload authorization.  Still
need to handle the "Overwrite" option on upload properly, i detect it now, but
end up rejecting the upload when it occurs.

--HG--
branch : Rigel