[Ticket #11211] - The ImageServer shall support upload and download operations through a REST API
Re-implemented a simpler version of the download.  When requesting a download,
we now return an array of URLs to download each of the sops in the
requested study/series/image.  Also created a test DownloadRestClient to actually
perform the file download and save to a specified directory.

Found a small problem in testing with the cancellation not working for
authorization requests.  Also adujusted the Bridge classes for authorization
to return an array of URLs in some cases.

--HG--
branch : Rigel