[Ticket #11211] - The ImageServer shall support upload and download operations through a REST API
Implementation for uploading/downloading files.  You can now upload a single
file through the rest API, and download the file using its associated UIDs.
The upload/download saves the file into a Raw Data SOP Class object in some
private attributes that are defined in the RawDataModule in the DICOM libary.
The routines also take into account if padding needed to be added to the file
(if the original file was odd length) and removes the padding if necessary). Note
that the NewSop notification was updated to include the original filename for the
file, and fixed an issue with the notification and telling if a sop has been
deidentified.

Also made some minor improvements to the test app integrate in the log4net logging
into the status windw and added upload/download file support.

--HG--
branch : Rigel