[Ticket #11148] - The ImageServer shall support a WCF API for configuration of research partitions
Added new RequestId parameter into the Configuration & Work requests.  When present
in the request, we search the database and check to make sure an identical request
was not already made.  If it has been made, an exception will occur.  The requests
also have a "DeletionTime" for configuration and work requests that can be set
as to when the request should "expire" and no longer be checked.  This time
would be in "localtime".

Note this change includes a database update to the ExternalRequestQueue table
adding a RequestId column (with an index on it).

Also as part of these updates, got the "state" notification for asynchronous
requests partially working.  it could still use some more work, but the basics
are operational for now.

--HG--
branch : Rigel