[Ticket #11263] - DICOM Library UID Generation Routine Flawed
Updated the UID generation mechanism to be based on cp-1156.  We now use
GUID based UIDs, which will be unique across app domains, contrary to the prior
scheme for UIDs.  Kept the old code around in the source file for reference
purposes, but it can't be accessed by the code.

--HG--
branch : Rigel