[Ticket #11209] - Research PACS notified of new arriving studies
Updated the NewSopNotification message to include a DicomFileLength and
ContentLength field.  ContentLength is filled in as the raw file length
for uploaded files or the PDF size for an encapsulated PDF file.

Also made some updates so when requesting Authorization for an upload/download,
a REST based state notification can be sent to the client.  This is specifically
for uploadfile/upload and it will include the total number of files uploaded.
Note that multiple state updates may be sent.  After the Upload request is
canceled, a final state with the ExternalRequestState field set to Complete will
be sent that should have the final count of objects uploaded.

--HG--
branch : Rigel