[Ticket #11350] - Create a rules engine for the RPACS
This commit contains some preliminary work to make the XmlSpecificationCompiler and XmlActionCompiler a little more general. Instead of defining global extension points by which to extend the compiler operators, the compiler classes now accept an extension point as a constructor argument (or optionally even a plain list of operators can be passed). This allows for the compiler classes to be reused in different scenarios where the same set of operators is not shared, within the same process (whereas with the old design, the global extension points would force any compiler instances within the same process to use the same set of operators).

A nice secondary effect of this change is that the 2nd generic parameter to IXmlActionCompilerOperator, TSchemaContext, has been eliminated, since there is no need to parametrize the schema when the set of operators itself is parameterizable.

In order to complete this refactoring, a number of changes had to be made to the ImageServer rules engine and the Dicom library, but these changes should have no functional impact.
 * The Dicom.Utilities.Rule class is now a little more general; it no longer has a TTypeEnum parameter, but instead requires compiler instances to be passed to its methods.
 * ServerRulesEngine now defines its own extension point for action operators - ServerRuleActionCompilerOperatorExtensionPoint. There is also an extension point for specification operators - DicomRuleSpecificationCompilerOperatorExtensionPoint - but it is located in the Dicom assembly because its extensions are there. Perhaps they could all be moved.
 * All of the IXmlActionCompilerOperator extensions in the ImageServer now require an additional attribute [ActionApplicability] to specify which type of rule(s) they are applicable for. The one exception is the NoOpActionOperator, which is applicable for all rule types.

--HG--
branch : Rigel