[Ticket #11352] - Enhanced US Volume support in the ImageServer [Ticket #11358] - Breast Tomo incorrectly set as Non-Image in IS Database
Discovered that the Enhanced US Volume SOP Class was inserted into the
ImageServer as a Non-Image SOP Class.  Updated the upgrade scripts to take
this into account.

Also, fixed the same issue with the Breast Tomo support (that this problem
originated with after copy/pasting the upgrade scripts from Brest Tomo).

Finally, made a small change to a test app.

--HG--
branch : Bennu