[Requirement #11393] - Ensure the source code repository is managed correctly [Ticket #11350] - Create a rules engine for the RPACS
Cleaned up an issue with the schemas for the rules engine, the recent
string comparisions needed seperate methods to define the scheme for each.
Added a shared method that just could be customized for each type of string
comparison.

Also fixed a small "Copy local" issue that was introduced with the imageserver's
shredhost, where it was copying some assemblies into the root directory.

--HG--
branch : Rigel