[Ticket #11350] - Create a rules engine for the RPACS
Since we cannot easily execute a process from IIS, the web process will have to ask the ShredHost process to execute the external application. Decided to add a NamedPipeConfiguration in case we want to  use that instead of TCP. However, in testing it, it doesn't seem to work.  Not sure what the problem is - don't have time to investigate.  Leaving the code in for future use, with comments in the code indicating that it doesn't currently work.

--HG--
branch : Rigel