[Ticket #11504] - The Research PACS shall be notified of changed/updated SOP Instances within studies
Added new UpdateSopEventArgs event for when a specific SOP instance is updated.
This could be doen during Edits, when processing reconcile or when processing
SOPs.  This event is fired when the situation occurs.

Also, introduced new UpdateSopNotification that is the same as the
NewSopNotification message, except with the different name.  The appropriate
changes were made for this.

Also, did a few things to update code.

--HG--
branch : Rigel