[Ticket #11504] - The Research PACS shall be notified of changed/updated SOP Instances within studies
Updates to make StudyEdits work with the changed SOP instance notification, as part of this:

Updated the StudyEditor to use the new event mechanism for firing events.  There
previously was a custom extension point for firing the events, changed so that we
have a new StudyEditedEventArgs event that is fired instead of the extension,
which fits in the new event mechanism.  There was an event fired before editing,
but it wasn't current used, so removed it.

--HG--
branch : Rigel