[Ticket #11500] - Accept Latest Duplicate SOP Policy for Research Partitions
Reworked the ImageServer handling of duplicate SOPs.  We now pass how to deal with
the SOPs through the WorkQueueUidData field,  The logic checking was done
previously at both the time of the import, and the time of the processing.
With this change, the primarily logic is now done at the time of import.

This allowed some simplifying of the logic in the DuplicateSopProcessorHelper.
Also moved the SopProcessingContext into its own source file, and renamed it
SopInstanceProcessorContext to more closely match where its used and so that
its named in a smilar way to the context used during import.

--HG--
branch : Rigel