[Ticket #11515] - Primary PACS Device Type
Introduced the PrimaryPacs DeviceType, and did the initial implementation.
Note that we don't update the Patient/Study/Series information properly yet
on the duplicates.  Also made updates to the context objects when importing
so that DuplicatePolicyEnum can be passed in to overwrite what would be the
normal duplicate policy.

Finally, fixed a small issue with the web.config file for Cleome, a reference
to 3ed was checked in.

--HG--
branch : Rigel