[Ticket #11515] - Primary PACS Device Type [Ticket #11500] - Accept Latest Duplicate SOP Policy for Research Partitions
Implemented the ability to update the Patient record a study is assigned to,
and the contents of the Study/Series/RequestAttributes tables, if applicable.
Made it so this update only occurs with the Primary PACS device type, over
overwrites just save the file and don't try and update the database tables.
Note that the AcceptLatest policy just overwrites the file, and does not
update counts.

--HG--
branch : Rigel