[Ticket #11211] - The ImageServer shall support upload and download operations through a REST API
Updated the Upload controllers to properly send State updates on upload.  We
now load the state from the database and update each time. The IsCancelable
property is set properly to false when we doing background processing so
the cancel can't be done until after processing is done.  Also updated the
test code to retry cancels.

--HG--
branch : Rigel