[Ticket #11294] - IntegrationServer Installer
Introduced new application root that can set a specific setting from the
command line in the enterprise settings store to a specific value.  The usage is as follows:

ClearCanvas.Executable.exe ClearCanvas.Enterprise.Common.Setup.SetSettingApplication
    /g:ClearCanvas.ImageServer.Common.EnterpriseImageServerServiceSettings
    /s:BaseUrl
    /v:http://SteveWinDesktop:9998
    /n:0.90

The /g option is requred ans sets the settings class, the /s option is required
and sets the specific setting in the settings class, the /v option sets the
value, and the /n option sets the version string for the setting.  The /n
option is optional, and if its not included, all the settings found that match
will be updated, independent of version string.

As part of the update, made changes to the build process to make the
ClearCanvas.Executable is built properly with the integration server.

--HG--
branch : Rigel