[Ticket #11146] - The ImageServer shall support research partitions
Updated the "Configure Server Partitions" page so that the Delete button is
disabled for all research partitions (usually the button is enabled for a
partition when there are no studies stored on the partition).

Also, updated the Edit Partition Dialog so that key fields that would break
research partition functionality are disabled in the dialog.  This includes
updating the AE title & port, the duplicate sop policy, and then the study
matching fields.

Updated the Studies page so that we disable deleting of studies on research
partitions.

--HG--
branch : Rigel