[Ticket #11503] - ImageServer partition authorization
Moved the PACS/Research/View Partitions token to AuthorityTokens.Research.ViewPartitions.

With this change, made it so the DataAccess.AllPartitions token is no longer
required for ImageDrive users.  Note that for now, the DataAccess.AllStudies
token is still required.  May think this through a bit further, but for now this
works.

--HG--
branch : Rigel