[Ticket #11560] - Default authority groups must be able to reference authority tokens from any component
Created a new project and solution called ClearCanvas.AuthorityGroups. The purpose of this project is to act as a central spot where we can define all the authority groups, hence allowing authority groups to reference tokens that are defined in different components.

 * The AuthorityGroups plugin defines a single application root: AuthorityGroupGenerator, which generates a set of XML files representing the set of default authority groups defined by IDefineAuthorityGroup extensions.
 * AuthorityGroupGenerator is intended to be run at packaging time, and then the appropriate output files can be collected and packaged into each installer.

TODO: need a way of importing authority group XML files remotely.

--HG--
branch : Rigel