[Ticket #11209] - Research PACS notified of newly arrived studies
This is from the rockstar issue tracking sheet, line 154.

When doing uploads, of multiple files, there was a concurrency issue with the
State objects that caused them to be overwritten.  Implemented a simple
offline concurrency check with a version column in the database.  Also
changed all the places where we update the state to first load it from the
database, then update it right away.  If a concurrency exception occurs,
we reload the row and reimplement the updates.  This seems to address the
issues.

As part of the changes, took the opportunity to implement a new class, ExternalRequestStateProxy
that handles the retries on concurrrency failure through the use of a delegate.
We may want to reduce the number of "state" object updates that are sent at some
point, but this is fine for now...

In any case, to add the new IS database column required w/ this changeset,
run this script:

ALTER TABLE dbo.ExternalRequestQueue ADD
	Revision int NOT NULL CONSTRAINT DF_ExternalRequestQueue_Revision DEFAULT 1
GO

--HG--
branch : Rigel