[Ticket #11523] - HL7 2.5.1 ORU^R01 Processor
Sprint polish.  Added in saving of the user account making a request. This
is automatically done in the bridge classes.  Also added in mechnaism to
save the client address of the system that make the request into the state
object.  This is also done automatically when the state object is created.

--HG--
branch : Rigel