[Ticket #11575] - Rules Engine support for testing tags within sequences
Introduced support for finding tag values like this:

$RequestAttributesSequence/$ScheduledProcedureStepId

If multiple instances of the specified tag are found in the message, the
match is done against a delimited tag value.  If, if the Scheduled Procedure Step Id
show up in two sequences with a value of A1234 and A5678, the resultant values
will be returned for testing a condition like this: "A1234\A5678".  In this
case, a regex test would match against one of the specific entries found if
attempting that.

--HG--
branch : Rigel