[Ticket #11211] - The ImageServer shall support upload and download operations
Introduced a new FailedSopEvent event into the core !ImageServer so that we can
get an event when a SOP fails processing.

Introduced into the NewSopNotification object a new ProcessingStatus field
that is an enum telling if the processing was successful or failed for a
SOP instance.  Note that the NewSopNotification may be sent multiple times
for the same sop with a failed status when a retry happens on a failed SOP.
The notification should only be sent 1 time with a Success status.

This is linked to line #280 in the Rockstar Issue Tracking database.

--HG--
branch : Rigel