[Ticket #11211] - The ImageServer shall support upload and download operations
More changes to handle notifications for uploads.  Decided to add a new event
into the imageserver for FailedUpdateSopEventArgs and renamed the FailedSopEventArgs
to FailedNewSopEventArgs to match up with the successfull update and new sop
events.  As part of this, updated all the processing locations so that there's
a matching failure event if a sop fails processing compared to if it succeeds.

Also reworked some of the event handlers in the enterprise class so that it
shares code for populating the events, which cleaned thm up a lot.

--HG--
branch : Rigel
rename : ImageServer/Core/Events/FailedSopEventArgs.cs => ImageServer/Core/Events/FailedNewSopEventArgs.cs