[Ticket #11523] - HL7 2.5.1 ORU^R01 Processor
Added changes to make a basic operating creator of encapsulated reports.  Need
to test more, but introduced a directory for templates to the web application,
and then added in the proper ReportPageModel to create the report and populate
the template.  Finally, reworked a bit how the report was created and put
it into a new ReportCreator class.

Needs some more testing and polish.

--HG--
branch : Rigel