[Requirement #11521] - ADT^A08 Patient Update Processor
Missed setting the Original Attributes SQ after edit, which was specified in
the requirement.  Added some general purpose code that this field will now be
updated whenever an edit occurs, which may also occur on reconcilliation.

Also, cleaned up some versioning of the ImageServer that were causing me
licensing issues and had to do an update to the configuration
files to setup the AuditService extension to be disabled.

--HG--
branch : Rigel