[Ticket #11602] - DBT Study Improvements for DICOM Transfers
Implemented saving incoming files directly to disk in the ImageServer.  As of
right now, only the Enhanced objects are saved this way, plus other multi-frame
objects.  The files are saved first to a temporary folder in
the <filesystem>\Partition\Streaming folder of the filesystem that the study is
stored on.  As part of the import, the file is then moved into the actual storage
folder.  Since its on the same filesystem, the move should be low cost operation.
The file should be cleaned up appropriately for the various duplicate SOP
policies and failure conditions.  Note, as of now, when a failure occurs writing
the file to disk, the association is aborted as opposed to the C-STORE operation
being failed.

Made basic updates to the RIS/ImageViewer + DICOM sample apps to take into account
low level changes in the interface for SCPs.

--HG--
branch : Rigel