[Ticket #11602] - DBT Study Improvements for DICOM Transfers
Implemented the ImageViewer DICOM receive code to be more efficient with
receiving DBT studies.  Was a lot easier than I thought.  Anyways, it only
does a "receive to disk" for the enhanced objects, and other multiframe objects.
Needs some more testing,and some polish to handle inserting a failed workitem
if we don't receive the object due to out of disk space condition.

--HG--
branch : Rigel