[Ticket #11602] - DBT Study Improvements for DICOM Transfers
Found a bug with the "streaming" implementation at CRL.  Specifically, when the
data contained in a P-DATA PDU coincided with the end of a tag's value, the
streaming code was stopping the parsing of the message received over the
network, even if we were not reading the message directly to disk.  We were
having problems with an Acuo implementation and an Intelerad implementation at
the customer site.

As part of debugging this issue, added some code for logging PDUs received.

--HG--
branch : Rigel