[Ticket #11208] - Objects stored in the VFS can be copied [Ticket #11528] - Objects stored in the VFS can be moved.
Removed the IRequestContext.Log and IRequestContext.LogOnCommit methods, and converted them to extension methods instead.  The purpose of the change is to avoid having to re-implement the methods whenever a new implementation of IRequestContext is created (there seem to be a lot of them).

Implemented a TransactionalLogMessageSink class that can hook into a System.Transaction and auto-flush on commit, or auto-purge on rollback. This makes the application code much simpler and saves everyone having to manually created log message sinks all over the place (and remember to flush them).

This commit is more of a general purpose refactoring as opposed to being specific to these TRs, but it needed to be done.

--HG--
branch : Rockstar