[Ticket #11678] ImageServer shredhost ignores service stop request if SQL server is not reachable
The primary cause of the problems was a the server alert that was being generated
at server startup.  It was stuck in a retry loop attempting to load the
AlertTypeEnums from the database at startup.  Moved this logging into a
Shred that simply logs the alert on startup/shutdown.

Also, made an adjustment to IPersistentStore to add a property, ShutdownRequested.
This is used by the Shreds to pass to any database connections that may be in a retry
mode to exit out of their retry loop and throw an exception.  This speeds up the
shutdown time.

Note that these fixes do not totaly address the issue, although its much better.
It still can take 6+ minutes for the ShredHostService to shut down, although
it makes more consistent progress when shutting down after the change, and it should
always shutdown.

--HG--
branch : Rigel