[Ticket #11673] QC reports are not accessible after re-inventory or re-apply server partition rules is run
Reworked based on code review how the rules engine is applied.  Updated the
StudyRulesEngine so that we can pass a ServerRulesEngine in the constructor.
This allows us to still cache the rules engine for each partition (so that the
database is not hit to load the rules each time they are applied).  The
StudyRulesEngine now is called to apply the rules.

Note that we no longer have to insert an ArchiveQueue entry as part of the
BaseReapplyRulesServiceLockItemProcessor, this is done as part fo the
StudyRulesEngine.

Finally, the previous solution for the bug would cause the database updates
for rules to be done twice, which is fixed.  Also, the StudyProccessed rules
were incorrectly applied for archived studies with the prior fix to the problem.

--HG--
branch : Rigel