[Ticket #11273] - Virtual File System shall be robust and have decent performance
Noticed while running Oto tests that a ChangeConflictException would often occur on saving Operation entities, and traced it to the SerializedData member which was being inadvertantly modified when it didn't need to be. Decided ultimately that the simplest way to fix this was to move the OperationToken out of the OperationState and store it in the serialized Request instead, because the Request is never modified.

Also, modified the UpdateContext class so that when we do get ChangeConflictException, we drill down and figure out exactly where the conflict is, and throw a more informative exception containing this information. This should make it easier to troubleshoot these kinds of issues in the future.

--HG--
branch : Rockstar