[Ticket #11747] - Integration items have lower priority in WorkQueue
Made updates to set the priority to Low in the WorkQueue for all external edits
and report creation.  Had to do some reworking of imports and inserting edits
to be able to have the option to set the priority on import.  Note that one
condition still sets the priority as High, when you import a Report that overwrites
a previous report.  I thought this case probably shoudl remain high, since data
is being updated, so that the latest info gets through.

--HG--
branch : Rigel