[Ticket #11750] Merge patients sometimes failed with stored procedure error
Issue was a "saved" value for the count of studies for the patient that was
later decrimented and may have been updated in the meantime.  Reworked the
way that the patient ids are handled to decrement the current value in a
database update, then check if the value is down to 0.

--HG--
branch : Rigel