[Requirement #11765] Webstation on ImageServer Web Components appears frozen and times out when loading large DBT study
Added some simple instrumentation for tomorrows build, will back out after testing.
This logs on the ImageServer side the time to read in and stream a file and on
the client side (in the web components) the time to receive the streamed data.
This will be logged in both the web components + workstation, so we can fiddle
a bit and monitor performance.

--HG--
branch : Rigel