[Ticket #11779] - ZIP archives uploaded through the Web GUI can be unpacked
In order to fix the issue with the INotificationQueueReader.Peek(..) implementation, needed to modify the PagedResults class and the way that paged result sets work:
 * Renamed PagedResults.NextPageToken to PositionToken, since it no longer indicates definitively whether there is another page or not.
 * The PageResults.HasNextPage member is now set explicitly, rather than using PositionToken == null to signal the condition. This has the advantage that we can retain the PositionToken value even if there are no more pages, since we now capture that information explicitly.

In the end however, decided to minimize the use of Peek(...) in favour of just adding some output parameters to convey whether there are more items in the queue, as Peek requires an additional database query.

Importing of ZIP archives now seems to work correctly, although it isn't particularly speedy. Doubt there's much that can be done about that though.

--HG--
branch : Rockstar