[Requirement #11763] DICOM, non-DICOM, and Study.xml files will be stored in Blob Storage
This is a working implementation where the RemoteBlobStore works over a WCF service.
We can now have the server read directly from the Stream on the client and read the
downloaded data and save it to disk.  Seems to work fairly well.  I still have the
services disabled at this point.  A few changes to the service providers were required
to make this work:

 * Added new ServiceTransferModeAttribute which allows you to set for a specific service contract that the TransferMode.Streaming can be used for a give ServiceContract
 * This sllows the standard service providers to be used to do the streaming.
 * Note that this works for NetTcp + NamedPipe binding, but not for WSHttp.

--HG--
branch : Rockstar